Update and rename Random-guess-number-thing-DS.py to Mod_6_Group_Project.py
Did another rename because realized that the file name needs to be the same as the main branch outline name just so merging can be more smoother. (thanks @Sledgehammer5568 for pointing that out!)
I also added header comments such as a small description and problems arising when testing out the code.